lcd4l: replace "\n" with string EOL
Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/b212d93b10f53eb58e049a2d81c07da3b3b717f1
Author: vanhofen <vanhofen@gmx.de>
Date: 2024-01-02 (Tue, 02 Jan 2024)

Origin message was:
------------------
- lcd4l: replace "\n" with string EOL

Conflicts:
	src/driver/lcd4l.cpp

The change seems to be an improvement in code quality, although it might not be strictly
necessary in this specific case, but from time to time, providing justifications in the
commit messages would also be a good idea, to be precise. That would be a nice thing to do.